PM-14445: Enable strict mode for KM team owned code
The aim was to enable `strictNullChecks` only, which will hopefully give better clarity on the types expectations from KM code, but to other teams and KM team. But since it is not possible to do that on individual module level (this flag is enabled everywhere in the code that KM module uses, resulting in 1000+ errors), the `typescript-strict-plugin` was used. As a consequence, the KM owned code is now in TS compiled strict mode.