rename new integration types to extension types
The first pass of the integration logic mixed names between
"integration" and "extension". For example, the remote service
has an "IntegrationId" whereas the part of the password manager
it extended has an "ExtensionPointId".

The new types have a more consistent approach. The remote
service now distinguishes the integrated product from the
vendor providing it, and the logic within the password manager
is now all the "extension". The specific location being extended
is now called the "site" or "extension site".

The main reason "extension" was chosen over "integration" is simply
so that the new and old types remain distinct in the folder hierarchy.